feat: Guard against a race between Start and Stop.

+++ pkg/jobmanager/job.go
@@ -159,6 +159,10 @@ func (j *concreteJob) Stop() error {
 		return nil
 	}
 
+	if j.cmd == nil || j.cmd.Process == nil {
+		return fmt.Errorf("job is in the running state but has not completed start")
+	}
+
 	if err := j.cmd.Process.Kill(); err != nil && err != os.ErrProcessDone {
 		return err
 	}
